preprocess: move EMSC to its own file
